Add/ward services

+++ public/components/ward/wards.controller.js
@@ -3,13 +3,12 @@
  */
 
 angular.module('inward').controller('WardController',
-    ['WardService', '$location', '$scope', 'ngNotify', 'sharedProperties','$mdDialog', function( WardService, $location, $scope, ngNotify, sharedProperties, $mdDialog) {
+    ['WardService', '$location', '$scope', 'ngNotify','$mdDialog','SweetAlert', function( WardService, $location, $scope, ngNotify, $mdDialog, SweetAlert) {
     var vm = this;
 
     function getWards() {
 
         WardService.get().then(wards =>{
-            console.log(wards);
             vm.wards = wards;
         })
     }
@@ -21,16 +20,30 @@ angular.module('inward').controller('WardController',
 
     $scope.deleteWard = (id) =>{
         "use strict";
-    
-        WardService.delete(id).then(()=>{
-            ngNotify.set('Ward Deleted!','error');
-            getWards();
-        })
+        SweetAlert.swal({
+            title: "Are you sure you want to delete?", 
+            text: "You will not be able to recover this record!", 
+            type: "warning", 
+            showCancelButton: true, 
+            confirmButtonColor: "#DD6B55",
+            confirmButtonText: "Delete",
+            closeOnConfirm: false,
+            closeOnCancel: true
+        }, 
+        function(isConfirm){
+            if(isConfirm){
+                WardService.delete(id).then(()=>{
+                    ngNotify.set('Ward Deleted!','error');
+                    getWards();
+                })
+                SweetAlert.swal("Deleted!");
+            }
+        });
+        
     }
 
     $scope.goToAddBed = (id) => {
-        $location.path('/beds');
-        sharedProperties.setWardNo(id);
+        $location.url('/wards/' + id+ '/beds');
     }
 
     $scope.addWard = (ward) => {
